requirements.txt
qrcode
Pillow
PyObjC
